coroutines

+++ 0x01-python_async_function/2-measure_runtime.py
@@ -5,7 +5,7 @@ import asyncio
 wait_n = __import__('1-concurrent_coroutines').wait_n
 
 
-def measure_time(n: int, max_delay: float) -> float:
+def measure_time(n: int, max_delay: int) -> float:
     """return total_time for a single task"""
     start = time.perf_counter()
     asyncio.run(wait_n(n, max_delay))
